+++ jdbc-postgres/build.gradle
@@ -16,6 +16,7 @@ dependencies {
     testImplementation project(':jdbc').sourceSets.test.output
     testImplementation project(':executor').sourceSets.test.output
     testImplementation project(':storage-local')
+    testImplementation project(':indexer')
     testImplementation project(':tests')
     testImplementation("io.micronaut.validation:micronaut-validation") // PostgresServiceLivenessCoordinatorTest fail to init without that
 }
