+++ settings.gradle
@@ -20,6 +20,7 @@ include 'webserver'
 include 'executor'
 include 'scheduler'
 include 'worker'
+include 'indexer'
 
 include 'ui'
 include 'model'
